pyproject.toml
[tool.poetry]
name = "ekklesia-common"
version = "20.6.0"
description = "Common code for the Ekklesia e-democracy platform"
readme = "README.md"
keywords = ["ekklesia", "edemocracy", "evoting", "participation"]
license = "AGPL-3.0-or-later"
authors = ["Tobias dpausp <dpausp@posteo.de>"]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8"
]

[tool.poetry.dependencies]
python = "^3.8"
munch = "^2.5.0"
morepath = "^0.19"
"more.babel_i18n" = "^20.5.0"
"more.browser_session" = "^19.8.0"
pyyaml = "^5.3.1"
sqlalchemy = "^1.3.17"
sqlalchemy_utils = "^0.36.6"
eliot = "^1.12.0"
deform = "^2.0.8"
pyjade = "^4.0.0"
py-gfm = "^0.1.4"
requests_oauthlib = "^1.3.0"
"zope.sqlalchemy" = "^1.3"
case-conversion = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-cov = "^2.8.1"
pytest-instafail = "^0.4.1"
responses = "^0.10.15"
webtest = "^2.0.35"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
